:truck: refactor: rename prompts to chains
